feature: BE와 로그인 통합 & document 목록 통합
