+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import {fetchCurrentUserServerSide} from "@/libs/user";
+import {fetchCurrentUserServerSide} from "@/libs/server/user";
 import {redirect} from "next/navigation";
 
 const LOGIN_URL = process.env.NEXT_PUBLIC_LOGIN_URL;
